+++ agixt/Agent.py
@@ -16,7 +16,8 @@ from DB import (
 from Providers import Providers
 from Extensions import Extensions
 from Globals import getenv, DEFAULT_SETTINGS, DEFAULT_USER
-from MagicalAuth import get_user_id
+from MagicalAuth import get_user_id, is_agixt_admin
+from agixtsdk import AGiXTSDK
 from fastapi import HTTPException
 from datetime import datetime, timezone, timedelta
 import logging
@@ -86,6 +87,51 @@ def add_agent(agent_name, provider_settings=None, commands=None, user=DEFAULT_US
     session.commit()
     session.close()
     return {"message": f"Agent {agent_name} created."}
+
+
+def webhook_create_user(
+    api_key: str,
+    email: str,
+    role: str = "user",
+    agent_name: str = "",
+    settings: dict = {},
+    commands: dict = {},
+    training_urls: list = [],
+    github_repos: list = [],
+    ApiClient: AGiXTSDK = AGiXTSDK(),
+):
+    if not is_agixt_admin(email=email, api_key=api_key):
+        return {"error": "Access Denied"}, 403
+    session = get_session()
+    email = email.lower()
+    user_exists = session.query(User).filter_by(email=email).first()
+    if user_exists:
+        session.close()
+        return {"error": "User already exists"}, 400
+    admin = True if role.lower() == "admin" else False
+    user = User(
+        email=email,
+        admin=admin,
+        first_name="",
+        last_name="",
+    )
+    session.add(user)
+    session.commit()
+    session.close()
+    if agent_name != "" and agent_name is not None:
+        add_agent(
+            agent_name=agent_name,
+            provider_settings=settings,
+            commands=commands,
+            user=email,
+        )
+    if training_urls != []:
+        for url in training_urls:
+            ApiClient.learn_url(agent_name=agent_name, url=url)
+    if github_repos != []:
+        for repo in github_repos:
+            ApiClient.learn_github_repo(agent_name=agent_name, github_repo=repo)
+    return {"status": "Success"}, 200
 
 
 def delete_agent(agent_name, user=DEFAULT_USER):
@@ -172,7 +218,7 @@ def get_agents(user=DEFAULT_USER):
 
 
 class Agent:
-    def __init__(self, agent_name=None, user=DEFAULT_USER, ApiClient=None):
+    def __init__(self, agent_name=None, user=DEFAULT_USER, ApiClient: AGiXTSDK = None):
         self.agent_name = agent_name if agent_name is not None else "AGiXT"
         user = user if user is not None else DEFAULT_USER
         self.user = user.lower()
